ACS Estimates by Census Block Group

Generate ACS 5yr estimates for multiple tables by Census Block Groups in multiple counties

Usage

get_acs_blockgroup(
  state = "Washington",
  counties = c("King", "Kitsap", "Pierce", "Snohomish"),
  table.names,
  years
)

Arguments

state

A character string state name or abbreviation. Defaults to Washington.

counties

A character string or vector of counties. Defaults to PSRC counties.

table.names

A character string or vector of Census table codes.

years

A numeric value or vector of years. An ACS year equal or greater than 2010 to the latest available year.

Value

a tibble of acs estimates by tracts for selected table codes. Includes detailed variable names.

Author

Craig Helmann