Decennial Estimates by Tract or Block Group and County

Generate decennial estimates for multiple tables by multiple tracts and/or counties

Usage

get_decennial_tract_county_bg(
  geography,
  counties = c("King", "Kitsap", "Pierce", "Snohomish"),
  variables,
  table_codes,
  years,
  state = "WA",
  sumfile
)

Arguments

geography

A character string as either 'tract', 'county', 'block group'.

counties

A character string or vector of counties. Defaults to PSRC counties.

variables

A character string or vector of Census variables

table_codes

A character string or vector of Census table codes.

years

Numeric or a vector of numeric years. A decennial year or years equal or greater than 2000.

state

A character string state abbreviation

sumfile

A character string for which summary file to use such as "sf1" or "dp"

Value

a tibble of decennial estimates by either tracts in a county/counties for selected table codes and years. Does not include variable names.

Author

Christy Lam