Retrieve and assemble PUMS data

The primary PUMS function

get_psrc_pums(
  span,
  dyear,
  level,
  vars,
  dollar_adj = TRUE,
  dir = NULL,
  labels = TRUE
)

Arguments

span
Either 1 for acs1 or 5 for acs5

dyear
The data year

level
Either "p" or "h", for "persons" or "households" respectively

vars
PUMS variable/s as an UPPERCASE string element or list

dollar_adj
Default TRUE; adjust income and cost values for inflation using the averaged factors provided with PUMS

dir
Directory for .gz file, if already downloaded. Default NULL uses the Census ftp.

labels
Default TRUE, recodes varible values to the corresponding label

Value

A srvyr object with appropriate sampling weight and replication weights

Examples

if (FALSE) {
Sys.getenv("CENSUS_API_KEY")}
get_psrc_pums(span=1, dyear=2019, level="p", vars=c("AGEP","SEX"))

#> Call: Called via srvyr
#> with 80 replicates and MSE variances.
#> Sampling variables:
#>  - repweights: `PWGTP1 + PWGTP2 + PWGTP3 + PWGTP4 + PWGTP5 + PWGTP6 + PWGTP7 + PWGTP8 + PWGTP9 + PWGTP10 + PWGTP11 + PWGTP12 + PWGTP13 + PWGTP14 + PWGTP15 + PWGTP16 + PWGTP17 + PWGTP18 + PWGTP19 + PWGTP20 + PWGTP21 + PWGTP22 + PWGTP23 + PWGTP24 + PWGTP25 + PWGTP26 + PWGTP27 + PWGTP28 + PWGTP29 + PWGTP30 + PWGTP31 + PWGTP32 + PWGTP33 + PWGTP34 + PWGTP35 + PWGTP36 + PWGTP37 + PWGTP38 + PWGTP39 + PWGTP40 + PWGTP41 + PWGTP42 + PWGTP43 + PWGTP44 + PWGTP45 + PWGTP46 + PWGTP47 + PWGTP48 + PWGTP49 + PWGTP50 + PWGTP51 + \n    PWGTP52 + PWGTP53 + PWGTP54 + PWGTP55 + PWGTP56 + PWGTP57 + PWGTP58 + PWGTP59 + PWGTP60 + PWGTP61 + PWGTP62 + PWGTP63 + PWGTP64 + PWGTP65 + PWGTP66 + PWGTP67 + PWGTP68 + PWGTP69 + PWGTP70 + PWGTP71 + PWGTP72 + PWGTP73 + PWGTP74 + PWGTP75 + PWGTP76 + PWGTP77 + PWGTP78 + PWGTP79 + PWGTP80`
#>  - weights: PWGTP
#> Data variables: SERIALNO (chr), SPORDER (fct), COUNTY (fct), AGEP (dbl), SEX
#>   (fct)