Format Estimates

Format estimates of a county summary table with the option to include shares and share margins of error.

format_estimates(table, type = "total", moe = TRUE)

Arguments

table
A data frame/tibble from `get_acs_recs()` for a single table and time period.

type
A character, select either 'total' or 'share'.

moe
A logical value, TRUE or FALSE to include or exclude margins of error.

Value

A data frame of ACS estimates by PSRC counties and region of either total estimates or proportions, with or without margins of error.

Author

Christy Lam

Examples

Sys.getenv("CENSUS_API_KEY")

#> [1] "c4780eb03010d73b7ae4e6894c1592375e545a21"

df <- psrccensus::get_acs_recs(geography = 'county',
                               table.names = c('B03002'),
                               years = c(2019),
                               acs.type = 'acs1')

#> The 1-year ACS provides data for geographies with populations of 65,000 and greater.

#> Getting data from the 2019 1-year ACS

format_estimates(df)

#> # A tibble: 21 x 15
#>    variable  label concept acs_type  year King_estimate King_moe Kitsap_estimate
#>    <chr>     <chr> <chr>   <chr>    <dbl>         <dbl>    <dbl>           <dbl>
#>  1 B03002_0~ Esti~ HISPAN~ acs1      2019       2252782        0          271473
#>  2 B03002_0~ Esti~ HISPAN~ acs1      2019       2030140        0          249096
#>  3 B03002_0~ Esti~ HISPAN~ acs1      2019       1302544     3208          205840
#>  4 B03002_0~ Esti~ HISPAN~ acs1      2019        147822     4678            7478
#>  5 B03002_0~ Esti~ HISPAN~ acs1      2019         13321     1990            2799
#>  6 B03002_0~ Esti~ HISPAN~ acs1      2019        424590     7085           13428
#>  7 B03002_0~ Esti~ HISPAN~ acs1      2019         15702     1831            2585
#>  8 B03002_0~ Esti~ HISPAN~ acs1      2019          6574     3281             276
#>  9 B03002_0~ Esti~ HISPAN~ acs1      2019        119587     8804           16690
#> 10 B03002_0~ Esti~ HISPAN~ acs1      2019          2639     1744             158
#> # ... with 11 more rows, and 7 more variables: Kitsap_moe <dbl>,
#> #   Pierce_estimate <dbl>, Pierce_moe <dbl>, Snohomish_estimate <dbl>,
#> #   Snohomish_moe <dbl>, Region_estimate <dbl>, Region_moe <dbl>

format_estimates(df, type = 'share', moe = FALSE)

#> # A tibble: 21 x 10
#>    variable   label  concept acs_type  year King_share Kitsap_share Pierce_share
#>    <chr>      <chr>  <chr>   <chr>    <dbl>      <dbl>        <dbl>        <dbl>
#>  1 B03002_001 Estim~ HISPAN~ acs1      2019    1           1            1       
#>  2 B03002_002 Estim~ HISPAN~ acs1      2019    0.901       0.918        0.886   
#>  3 B03002_003 Estim~ HISPAN~ acs1      2019    0.578       0.758        0.656   
#>  4 B03002_004 Estim~ HISPAN~ acs1      2019    0.0656      0.0275       0.0651  
#>  5 B03002_005 Estim~ HISPAN~ acs1      2019    0.00591     0.0103       0.00925 
#>  6 B03002_006 Estim~ HISPAN~ acs1      2019    0.188       0.0495       0.0623  
#>  7 B03002_007 Estim~ HISPAN~ acs1      2019    0.00697     0.00952      0.0132  
#>  8 B03002_008 Estim~ HISPAN~ acs1      2019    0.00292     0.00102      0.00199 
#>  9 B03002_009 Estim~ HISPAN~ acs1      2019    0.0531      0.0615       0.0784  
#> 10 B03002_010 Estim~ HISPAN~ acs1      2019    0.00117     0.000582     0.000864
#> # ... with 11 more rows, and 2 more variables: Snohomish_share <dbl>,
#> #   Region_share <dbl>

t <- get_acs_recs(geography = 'county',
                  counties = c("King", "Kitsap"),
                  table.names = c('B03002'),
                  years = c(2019),
                  acs.type = 'acs1')

#> The 1-year ACS provides data for geographies with populations of 65,000 and greater.

#> Getting data from the 2019 1-year ACS

format_estimates(t, type = 'share', moe = TRUE)

#> # A tibble: 21 x 9
#>    variable  label concept acs_type  year King_share King_share_moe Kitsap_share
#>    <chr>     <chr> <chr>   <chr>    <dbl>      <dbl>          <dbl>        <dbl>
#>  1 B03002_0~ Esti~ HISPAN~ acs1      2019    1             0            1       
#>  2 B03002_0~ Esti~ HISPAN~ acs1      2019    0.901         0            0.918   
#>  3 B03002_0~ Esti~ HISPAN~ acs1      2019    0.578         0.00142      0.758   
#>  4 B03002_0~ Esti~ HISPAN~ acs1      2019    0.0656        0.00208      0.0275  
#>  5 B03002_0~ Esti~ HISPAN~ acs1      2019    0.00591       0.000883     0.0103  
#>  6 B03002_0~ Esti~ HISPAN~ acs1      2019    0.188         0.00315      0.0495  
#>  7 B03002_0~ Esti~ HISPAN~ acs1      2019    0.00697       0.000813     0.00952 
#>  8 B03002_0~ Esti~ HISPAN~ acs1      2019    0.00292       0.00146      0.00102 
#>  9 B03002_0~ Esti~ HISPAN~ acs1      2019    0.0531        0.00391      0.0615  
#> 10 B03002_0~ Esti~ HISPAN~ acs1      2019    0.00117       0.000774     0.000582
#> # ... with 11 more rows, and 1 more variable: Kitsap_share_moe <dbl>