ACS Estimates

Generate ACS estimates for multiple tables by tracts, counties, MSAs, or places for multiple years.

get_acs_recs(
  geography,
  state = "Washington",
  counties = c("King", "Kitsap", "Pierce", "Snohomish"),
  table.names,
  years,
  FIPS = c("14740", "42660"),
  acs.type
)

Arguments

geography
A character string as either 'tract', 'county', 'msa', or 'place'.

state
A character string state name or abbreviation. Defaults to Washington.

counties
A character string or vector of counties. Defaults to PSRC counties.

table.names
A character string or vector of Census table codes.

years
A numeric value or vector of years. An ACS year equal or greater than 2010 to the latest available year.

FIPS
Character string for FIPS codes for specific MSA geographies. Defaults to Seattle & Bremerton MSA c("14740","42660")

acs.type
A character string as either 'acs1', 'acs3' or acs5'.

Value

a tibble of ACS estimates by selected geography for selected table codes. Includes variable names.

Author

Craig Helmann

Examples

if (FALSE) {
Sys.getenv("CENSUS_API_KEY")}
get_acs_recs(geography = 'county',
             table.names = c('B03002',"C17002"),
             years=c(2017,2019),
             acs.type = 'acs1')

#> The 1-year ACS provides data for geographies with populations of 65,000 and greater.

#> Getting data from the 2017 1-year ACS

#> Loading ACS1 variables for 2017 from table B03002. To cache this dataset for faster access to ACS tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per ACS dataset.

#> The 1-year ACS provides data for geographies with populations of 65,000 and greater.

#> Getting data from the 2019 1-year ACS

#> The 1-year ACS provides data for geographies with populations of 65,000 and greater.

#> Getting data from the 2017 1-year ACS

#> Loading ACS1 variables for 2017 from table C17002. To cache this dataset for faster access to ACS tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per ACS dataset.

#> The 1-year ACS provides data for geographies with populations of 65,000 and greater.

#> Getting data from the 2019 1-year ACS

#> # A tibble: 290 x 11
#>    GEOID name   state variable estimate   moe label    concept  census_geography
#>    <chr> <chr>  <chr> <chr>       <dbl> <dbl> <chr>    <chr>    <chr>           
#>  1 53033 King ~ Wash~ B03002_~  2188649    NA Estimat~ HISPANI~ County          
#>  2 53033 King ~ Wash~ B03002_~  1976237    NA Estimat~ HISPANI~ County          
#>  3 53033 King ~ Wash~ B03002_~  1309568  1385 Estimat~ HISPANI~ County          
#>  4 53033 King ~ Wash~ B03002_~   133498  5330 Estimat~ HISPANI~ County          
#>  5 53033 King ~ Wash~ B03002_~     8058  1683 Estimat~ HISPANI~ County          
#>  6 53033 King ~ Wash~ B03002_~   379302  5696 Estimat~ HISPANI~ County          
#>  7 53033 King ~ Wash~ B03002_~    16765  1987 Estimat~ HISPANI~ County          
#>  8 53033 King ~ Wash~ B03002_~     7055  2296 Estimat~ HISPANI~ County          
#>  9 53033 King ~ Wash~ B03002_~   121991  8954 Estimat~ HISPANI~ County          
#> 10 53033 King ~ Wash~ B03002_~     3853  1480 Estimat~ HISPANI~ County          
#> # ... with 280 more rows, and 2 more variables: acs_type <chr>, year <dbl>

get_acs_recs(geography = 'county',
             table.names = c('B03002'),
             counties=c("Kitsap"),
             years=c(2019),
             acs.type = 'acs5')

#> Getting data from the 2015-2019 5-year ACS

#> # A tibble: 21 x 11
#>    GEOID name   state variable estimate   moe label    concept  census_geography
#>    <chr> <chr>  <chr> <chr>       <dbl> <dbl> <chr>    <chr>    <chr>           
#>  1 53035 Kitsa~ Wash~ B03002_~   265882    NA Estimat~ HISPANI~ County          
#>  2 53035 Kitsa~ Wash~ B03002_~   245230    NA Estimat~ HISPANI~ County          
#>  3 53035 Kitsa~ Wash~ B03002_~   203899   101 Estimat~ HISPANI~ County          
#>  4 53035 Kitsa~ Wash~ B03002_~     6821   491 Estimat~ HISPANI~ County          
#>  5 53035 Kitsa~ Wash~ B03002_~     2255   299 Estimat~ HISPANI~ County          
#>  6 53035 Kitsa~ Wash~ B03002_~    12580   660 Estimat~ HISPANI~ County          
#>  7 53035 Kitsa~ Wash~ B03002_~     2128   203 Estimat~ HISPANI~ County          
#>  8 53035 Kitsa~ Wash~ B03002_~      205    98 Estimat~ HISPANI~ County          
#>  9 53035 Kitsa~ Wash~ B03002_~    17342   899 Estimat~ HISPANI~ County          
#> 10 53035 Kitsa~ Wash~ B03002_~      170    72 Estimat~ HISPANI~ County          
#> # ... with 11 more rows, and 2 more variables: acs_type <chr>, year <dbl>

get_acs_recs(geography = 'msa',
             table.names = c('B03002',"C17002"),
             years=c(2019),
             acs.type = 'acs1')

#> The 1-year ACS provides data for geographies with populations of 65,000 and greater.

#> Getting data from the 2019 1-year ACS

#> No encoding supplied: defaulting to UTF-8.

#> Error: Your API call has errors.  The API message returned is Sorry, the system is currently undergoing maintenance or is busy.  Please try again later..

get_acs_recs(geography = 'msa',
             table.names = c("C17002"),
             years=c(2019),
             FIPS = c("14740"),
             acs.type = 'acs5')

#> Getting data from the 2015-2019 5-year ACS

#> # A tibble: 8 x 11
#>   GEOID name    state  variable estimate   moe label  concept   census_geography
#>   <chr> <chr>   <chr>  <chr>       <dbl> <dbl> <chr>  <chr>     <chr>           
#> 1 14740 Bremer~ WA Me~ C17002_~   257272   703 Estim~ RATIO OF~ MSA             
#> 2 14740 Bremer~ WA Me~ C17002_~    10948   991 Estim~ RATIO OF~ MSA             
#> 3 14740 Bremer~ WA Me~ C17002_~    11456  1149 Estim~ RATIO OF~ MSA             
#> 4 14740 Bremer~ WA Me~ C17002_~     8101  1178 Estim~ RATIO OF~ MSA             
#> 5 14740 Bremer~ WA Me~ C17002_~     6950   861 Estim~ RATIO OF~ MSA             
#> 6 14740 Bremer~ WA Me~ C17002_~    12207  1176 Estim~ RATIO OF~ MSA             
#> 7 14740 Bremer~ WA Me~ C17002_~     5582   936 Estim~ RATIO OF~ MSA             
#> 8 14740 Bremer~ WA Me~ C17002_~   202028  1990 Estim~ RATIO OF~ MSA             
#> # ... with 2 more variables: acs_type <chr>, year <dbl>

get_acs_recs(geography = 'place',
             table.names = c("C17002"),
             years=c(2019),
             acs.type = 'acs5')

#> Getting data from the 2015-2019 5-year ACS

#> No encoding supplied: defaulting to UTF-8.

#> Error: lexical error: invalid char in json text.
#>                                        Sorry, the system is currently 
#>                      (right here) ------^

get_acs_recs(geography = 'tract',
             counties=c("Pierce"),
             table.names = c('B03002'),
             years=c(2018,2019))

#> Getting data from the 2014-2018 5-year ACS

#> Getting data from the 2015-2019 5-year ACS

#> # A tibble: 7,224 x 11
#>    GEOID  name   state variable estimate   moe label   concept  census_geography
#>    <chr>  <chr>  <chr> <chr>       <dbl> <dbl> <chr>   <chr>    <chr>           
#>  1 53053~ Censu~ Wash~ B03002_~     2414   371 Estima~ HISPANI~ Tract           
#>  2 53053~ Censu~ Wash~ B03002_~     1388   222 Estima~ HISPANI~ Tract           
#>  3 53053~ Censu~ Wash~ B03002_~      873   147 Estima~ HISPANI~ Tract           
#>  4 53053~ Censu~ Wash~ B03002_~      155   120 Estima~ HISPANI~ Tract           
#>  5 53053~ Censu~ Wash~ B03002_~        8    12 Estima~ HISPANI~ Tract           
#>  6 53053~ Censu~ Wash~ B03002_~      289   131 Estima~ HISPANI~ Tract           
#>  7 53053~ Censu~ Wash~ B03002_~       18    41 Estima~ HISPANI~ Tract           
#>  8 53053~ Censu~ Wash~ B03002_~        5     8 Estima~ HISPANI~ Tract           
#>  9 53053~ Censu~ Wash~ B03002_~       40    46 Estima~ HISPANI~ Tract           
#> 10 53053~ Censu~ Wash~ B03002_~        0    12 Estima~ HISPANI~ Tract           
#> # ... with 7,214 more rows, and 2 more variables: acs_type <chr>, year <dbl>