ACS Estimates

Generate ACS estimates for multiple tables by tracts, counties, MSAs, or places for multiple years.

Usage

get_acs_recs(
  geography,
  state = "Washington",
  counties = c("King", "Kitsap", "Pierce", "Snohomish"),
  table.names,
  years,
  FIPS = c("14740", "42660"),
  place_FIPS = NULL,
  acs.type
)

Arguments

geography

A character string as either 'tract', 'county', 'msa', or 'place'.

state

A character string state name or abbreviation. Defaults to Washington.

counties

A character string or vector of counties. Defaults to PSRC counties.

table.names

A character string or vector of Census table codes.

years

A numeric value or vector of years. An ACS year equal or greater than 2010 to the latest available year.

FIPS

Character string for FIPS codes for specific MSA geographies. Defaults to Seattle & Bremerton MSA c("14740","42660")

place_FIPS

Character string of FIPS codes (with state prefix) for specific Census Places. If NULL, Places within the PSRC Region will be returned.

acs.type

A character string as either 'acs1', 'acs3' or acs5'.

Value

a tibble of ACS estimates by selected geography for selected table codes. Includes variable names.

Author

Craig Helmann

Examples

get_acs_recs(geography = 'county',
             table.names = c('B03002'),
             years=c(2017,2019),
             acs.type = 'acs1')
#> Getting data from the 2017 1-year ACS
#> The 1-year ACS provides data for geographies with populations of 65,000 and greater.
#> Loading ACS1 variables for 2017 from table B03002. To cache this dataset for faster access to ACS tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per ACS dataset.
#> Using FIPS code '53' for state 'Washington'
#> Using FIPS code '033' for 'King County'
#> Using FIPS code '035' for 'Kitsap County'
#> Using FIPS code '053' for 'Pierce County'
#> Using FIPS code '061' for 'Snohomish County'
#> Getting data from the 2019 1-year ACS
#> The 1-year ACS provides data for geographies with populations of 65,000 and greater.
#> Loading ACS1 variables for 2019 from table B03002. To cache this dataset for faster access to ACS tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per ACS dataset.
#> Using FIPS code '53' for state 'Washington'
#> Using FIPS code '033' for 'King County'
#> Using FIPS code '035' for 'Kitsap County'
#> Using FIPS code '053' for 'Pierce County'
#> Using FIPS code '061' for 'Snohomish County'
#> # A tibble: 210 × 14
#>    GEOID name       state variable estimate   moe label concept census_geography
#>    <chr> <chr>      <chr> <chr>       <dbl> <dbl> <chr> <chr>   <chr>           
#>  1 53033 King Coun… Wash… B03002_…  2188649    NA Esti… HISPAN… County          
#>  2 53033 King Coun… Wash… B03002_…  1976237    NA Esti… HISPAN… County          
#>  3 53033 King Coun… Wash… B03002_…  1309568  1385 Esti… HISPAN… County          
#>  4 53033 King Coun… Wash… B03002_…   133498  5330 Esti… HISPAN… County          
#>  5 53033 King Coun… Wash… B03002_…     8058  1683 Esti… HISPAN… County          
#>  6 53033 King Coun… Wash… B03002_…   379302  5696 Esti… HISPAN… County          
#>  7 53033 King Coun… Wash… B03002_…    16765  1987 Esti… HISPAN… County          
#>  8 53033 King Coun… Wash… B03002_…     7055  2296 Esti… HISPAN… County          
#>  9 53033 King Coun… Wash… B03002_…   121991  8954 Esti… HISPAN… County          
#> 10 53033 King Coun… Wash… B03002_…     3853  1480 Esti… HISPAN… County          
#> # ℹ 200 more rows
#> # ℹ 5 more variables: acs_type <chr>, year <dbl>, se <dbl>, cv <dbl>,
#> #   reliability <chr>

get_acs_recs(geography = 'county',
             table.names = c("C17002"),
             counties=c("Kitsap"),
             years=c(2019),
             acs.type = 'acs5')
#> Getting data from the 2015-2019 5-year ACS
#> Using FIPS code '53' for state 'Washington'
#> Using FIPS code '035' for 'Kitsap County'
#> # A tibble: 8 × 14
#>   GEOID name        state variable estimate   moe label concept census_geography
#>   <chr> <chr>       <chr> <chr>       <dbl> <dbl> <chr> <chr>   <chr>           
#> 1 53035 Kitsap Cou… Wash… C17002_…   257272   703 Esti… RATIO … County          
#> 2 53035 Kitsap Cou… Wash… C17002_…    10948   991 Esti… RATIO … County          
#> 3 53035 Kitsap Cou… Wash… C17002_…    11456  1149 Esti… RATIO … County          
#> 4 53035 Kitsap Cou… Wash… C17002_…     8101  1178 Esti… RATIO … County          
#> 5 53035 Kitsap Cou… Wash… C17002_…     6950   861 Esti… RATIO … County          
#> 6 53035 Kitsap Cou… Wash… C17002_…    12207  1176 Esti… RATIO … County          
#> 7 53035 Kitsap Cou… Wash… C17002_…     5582   936 Esti… RATIO … County          
#> 8 53035 Kitsap Cou… Wash… C17002_…   202028  1990 Esti… RATIO … County          
#> # ℹ 5 more variables: acs_type <chr>, year <dbl>, se <dbl>, cv <dbl>,
#> #   reliability <chr>