Decennial Estimates

Generate decennial estimates for multiple tables by tracts, counties, MSAs, or places. Currently only working for SF1 tables.

Usage

get_decennial_recs(
  geography,
  counties = c("King", "Kitsap", "Pierce", "Snohomish"),
  table_codes,
  years,
  fips = NULL
)

Arguments

geography

A character string as either 'tract', 'county', 'block group', 'msa', or 'place'.

counties

A character string or vector of counties. Defaults to PSRC counties.

table_codes

A character string or vector of Census table codes, the table code will be padded with 0s such as "H001", as opposed to "H1

years

Numeric or a vector of numeric years. A decennial year or years equal or greater than 2000.

fips

Character. Single code or vector of either MSA or place fips codes.

Value

a tibble of decennial estimates by selected geography for selected table codes. Includes variable names.

Author

Christy Lam

Examples

tbl_names <- paste0('PCT020', LETTERS[1:6])
get_decennial_recs(geography = 'county', table_codes = tbl_names, years = 2010)
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table PCT020A. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '033' for 'King County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table PCT020A. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '035' for 'Kitsap County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table PCT020A. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '053' for 'Pierce County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table PCT020A. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '061' for 'Snohomish County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table PCT020B. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '033' for 'King County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table PCT020B. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '035' for 'Kitsap County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table PCT020B. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '053' for 'Pierce County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table PCT020B. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '061' for 'Snohomish County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table PCT020C. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '033' for 'King County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table PCT020C. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '035' for 'Kitsap County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table PCT020C. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '053' for 'Pierce County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table PCT020C. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '061' for 'Snohomish County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table PCT020D. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '033' for 'King County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table PCT020D. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '035' for 'Kitsap County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table PCT020D. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '053' for 'Pierce County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table PCT020D. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '061' for 'Snohomish County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table PCT020E. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '033' for 'King County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table PCT020E. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '035' for 'Kitsap County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table PCT020E. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '053' for 'Pierce County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table PCT020E. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '061' for 'Snohomish County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table PCT020F. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '033' for 'King County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table PCT020F. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '035' for 'Kitsap County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table PCT020F. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '053' for 'Pierce County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table PCT020F. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '061' for 'Snohomish County'
#> Using Census Summary File 1
#> # A tibble: 960 × 7
#>    GEOID NAME                    variable   value  year label            concept
#>    <chr> <chr>                   <chr>      <dbl> <dbl> <chr>            <chr>  
#>  1 53033 King County, Washington PCT020A001 24629  2010 Total            GROUP …
#>  2 53033 King County, Washington PCT020A002  7394  2010 Total!!Institut… GROUP …
#>  3 53033 King County, Washington PCT020A003  1977  2010 Total!!Institut… GROUP …
#>  4 53033 King County, Washington PCT020A004   392  2010 Total!!Institut… GROUP …
#>  5 53033 King County, Washington PCT020A005     0  2010 Total!!Institut… GROUP …
#>  6 53033 King County, Washington PCT020A006     0  2010 Total!!Institut… GROUP …
#>  7 53033 King County, Washington PCT020A007  1287  2010 Total!!Institut… GROUP …
#>  8 53033 King County, Washington PCT020A008   298  2010 Total!!Institut… GROUP …
#>  9 53033 King County, Washington PCT020A009     0  2010 Total!!Institut… GROUP …
#> 10 53033 King County, Washington PCT020A010   251  2010 Total!!Institut… GROUP …
#> # ℹ 950 more rows

get_decennial_recs(geography = 'county', table_codes = 'P001', years = c(2000, 2010))
#> Getting data from the 2000 decennial Census
#> Loading SF1 variables for 2000 from table P001. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '033' for 'King County'
#> Using Census Summary File 1
#> Getting data from the 2000 decennial Census
#> Loading SF1 variables for 2000 from table P001. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '035' for 'Kitsap County'
#> Using Census Summary File 1
#> Getting data from the 2000 decennial Census
#> Loading SF1 variables for 2000 from table P001. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '053' for 'Pierce County'
#> Using Census Summary File 1
#> Getting data from the 2000 decennial Census
#> Loading SF1 variables for 2000 from table P001. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '061' for 'Snohomish County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table P001. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '033' for 'King County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table P001. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '035' for 'Kitsap County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table P001. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '053' for 'Pierce County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table P001. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '061' for 'Snohomish County'
#> Using Census Summary File 1
#> 
#> Concept for table codes may differ across Census years. Please double check with tidycensus::load_variables()
#> # A tibble: 10 × 7
#>    GEOID  NAME                         variable   value  year label concept     
#>    <chr>  <chr>                        <chr>      <dbl> <dbl> <chr> <chr>       
#>  1 53033  King County, Washington      P001001  1737034  2000 Total TOTAL POPUL…
#>  2 53035  Kitsap County, Washington    P001001   231969  2000 Total TOTAL POPUL…
#>  3 53053  Pierce County, Washington    P001001   700820  2000 Total TOTAL POPUL…
#>  4 53061  Snohomish County, Washington P001001   606024  2000 Total TOTAL POPUL…
#>  5 REGION Region                       P001001  3275847  2000 Total TOTAL POPUL…
#>  6 53033  King County, Washington      P001001  1931249  2010 Total TOTAL POPUL…
#>  7 53035  Kitsap County, Washington    P001001   251133  2010 Total TOTAL POPUL…
#>  8 53053  Pierce County, Washington    P001001   795225  2010 Total TOTAL POPUL…
#>  9 53061  Snohomish County, Washington P001001   713335  2010 Total TOTAL POPUL…
#> 10 REGION Region                       P001001  3690942  2010 Total TOTAL POPUL…

get_decennial_recs(geography = 'tract', table_codes = tbl_names, years = 2010)
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table PCT020A. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '033' for 'King County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table PCT020A. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '035' for 'Kitsap County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table PCT020A. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '053' for 'Pierce County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table PCT020A. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '061' for 'Snohomish County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table PCT020B. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '033' for 'King County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table PCT020B. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '035' for 'Kitsap County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table PCT020B. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '053' for 'Pierce County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table PCT020B. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '061' for 'Snohomish County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table PCT020C. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '033' for 'King County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table PCT020C. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '035' for 'Kitsap County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table PCT020C. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '053' for 'Pierce County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table PCT020C. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '061' for 'Snohomish County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table PCT020D. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '033' for 'King County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table PCT020D. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '035' for 'Kitsap County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table PCT020D. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '053' for 'Pierce County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table PCT020D. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '061' for 'Snohomish County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table PCT020E. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '033' for 'King County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table PCT020E. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '035' for 'Kitsap County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table PCT020E. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '053' for 'Pierce County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table PCT020E. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '061' for 'Snohomish County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table PCT020F. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '033' for 'King County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table PCT020F. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '035' for 'Kitsap County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table PCT020F. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '053' for 'Pierce County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table PCT020F. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '061' for 'Snohomish County'
#> Using Census Summary File 1
#> # A tibble: 148,992 × 7
#>    GEOID       NAME                           variable value  year label concept
#>    <chr>       <chr>                          <chr>    <dbl> <dbl> <chr> <chr>  
#>  1 53033000100 Census Tract 1, King County, … PCT020A…    32  2010 Total GROUP …
#>  2 53033000200 Census Tract 2, King County, … PCT020A…    55  2010 Total GROUP …
#>  3 53033000300 Census Tract 3, King County, … PCT020A…     8  2010 Total GROUP …
#>  4 53033000401 Census Tract 4.01, King Count… PCT020A…    97  2010 Total GROUP …
#>  5 53033000402 Census Tract 4.02, King Count… PCT020A…   247  2010 Total GROUP …
#>  6 53033000500 Census Tract 5, King County, … PCT020A…    17  2010 Total GROUP …
#>  7 53033000600 Census Tract 6, King County, … PCT020A…    34  2010 Total GROUP …
#>  8 53033000700 Census Tract 7, King County, … PCT020A…     9  2010 Total GROUP …
#>  9 53033000800 Census Tract 8, King County, … PCT020A…    52  2010 Total GROUP …
#> 10 53033000900 Census Tract 9, King County, … PCT020A…    13  2010 Total GROUP …
#> # ℹ 148,982 more rows

get_decennial_recs(geography = 'place',
                   table_codes = 'PCT013',
                   years = 2010,
                   fips = c("5363000", "5308850"))
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table PCT013. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using Census Summary File 1
#> Using FIPS code '53' for state 'WA'
#> Using Census Summary File 1
#> # A tibble: 98 × 7
#>    GEOID   NAME                     variable   value  year label         concept
#>    <chr>   <chr>                    <chr>      <dbl> <dbl> <chr>         <chr>  
#>  1 5308850 Burien city, Washington  PCT013001  33013  2010 Total         SEX BY…
#>  2 5363000 Seattle city, Washington PCT013001 583735  2010 Total         SEX BY…
#>  3 5308850 Burien city, Washington  PCT013002  16612  2010 Total!!Male   SEX BY…
#>  4 5363000 Seattle city, Washington PCT013002 290243  2010 Total!!Male   SEX BY…
#>  5 5308850 Burien city, Washington  PCT013003   1140  2010 Total!!Male!… SEX BY…
#>  6 5363000 Seattle city, Washington PCT013003  16424  2010 Total!!Male!… SEX BY…
#>  7 5308850 Burien city, Washington  PCT013004   1065  2010 Total!!Male!… SEX BY…
#>  8 5363000 Seattle city, Washington PCT013004  12919  2010 Total!!Male!… SEX BY…
#>  9 5308850 Burien city, Washington  PCT013005    947  2010 Total!!Male!… SEX BY…
#> 10 5363000 Seattle city, Washington PCT013005  10989  2010 Total!!Male!… SEX BY…
#> # ℹ 88 more rows

get_decennial_recs(geography = 'msa',
                   table_codes = c("H001", "P001"),
                   years = c(2000, 2010),
                   fips = c('42660', "28420"))
#> Getting data from the 2000 decennial Census
#> Loading SF1 variables for 2000 from table H001. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Error : Your API call has errors.  The API message returned is error: unknown/unsupported geography heirarchy.
#> Variables not found in Summary File 1. Trying Summary File 3...
#> [1] "API error, the year 2000 requested may not be available."
#> Getting data from the 2000 decennial Census
#> Loading SF1 variables for 2000 from table P001. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Error : Your API call has errors.  The API message returned is error: unknown/unsupported geography heirarchy.
#> Variables not found in Summary File 1. Trying Summary File 3...
#> [1] "API error, the year 2000 requested may not be available."
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table H001. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table P001. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using Census Summary File 1
#> 
#> Concept for table codes may differ across Census years. Please double check with tidycensus::load_variables()
#> # A tibble: 4 × 7
#>   GEOID NAME                                 variable  value  year label concept
#>   <chr> <chr>                                <chr>     <dbl> <dbl> <chr> <chr>  
#> 1 28420 Kennewick-Pasco-Richland, WA Metro … H001001  9.30e4  2010 Total HOUSIN…
#> 2 42660 Seattle-Tacoma-Bellevue, WA Metro A… H001001  1.46e6  2010 Total HOUSIN…
#> 3 28420 Kennewick-Pasco-Richland, WA Metro … P001001  2.53e5  2010 Total TOTAL …
#> 4 42660 Seattle-Tacoma-Bellevue, WA Metro A… P001001  3.44e6  2010 Total TOTAL …

get_decennial_recs(geography = 'block group',
                   table_codes = c('H001', 'H006'),
                   years = 2010)
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table H001. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '033' for 'King County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table H001. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '035' for 'Kitsap County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table H001. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '053' for 'Pierce County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table H001. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '061' for 'Snohomish County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table H006. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '033' for 'King County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table H006. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '035' for 'Kitsap County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table H006. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '053' for 'Pierce County'
#> Using Census Summary File 1
#> Getting data from the 2010 decennial Census
#> Loading SF1 variables for 2010 from table H006. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.
#> Using FIPS code '53' for state 'WA'
#> Using FIPS code '061' for 'Snohomish County'
#> Using Census Summary File 1
#> # A tibble: 23,823 × 7
#>    GEOID        NAME                          variable value  year label concept
#>    <chr>        <chr>                         <chr>    <dbl> <dbl> <chr> <chr>  
#>  1 530330001001 Block Group 1, Census Tract … H001001    565  2010 Total HOUSIN…
#>  2 530330001002 Block Group 2, Census Tract … H001001    658  2010 Total HOUSIN…
#>  3 530330001003 Block Group 3, Census Tract … H001001    744  2010 Total HOUSIN…
#>  4 530330001004 Block Group 4, Census Tract … H001001    838  2010 Total HOUSIN…
#>  5 530330001005 Block Group 5, Census Tract … H001001    638  2010 Total HOUSIN…
#>  6 530330002001 Block Group 1, Census Tract … H001001    455  2010 Total HOUSIN…
#>  7 530330002002 Block Group 2, Census Tract … H001001    590  2010 Total HOUSIN…
#>  8 530330002003 Block Group 3, Census Tract … H001001    826  2010 Total HOUSIN…
#>  9 530330002004 Block Group 4, Census Tract … H001001    576  2010 Total HOUSIN…
#> 10 530330002005 Block Group 5, Census Tract … H001001    698  2010 Total HOUSIN…
#> # ℹ 23,813 more rows