Get equity tracts

Retrieve a 0-1 variable for whether each tract in the region is in an equity geography or not

get_equity_geographies(
  equity_type = "50",
  equity_group = "race",
  year = "2019"
)

Arguments

equity_type
string, The type of equity geography either '50' (50 percent) or 'regavg' (regional average)

equity_group
string, The group of people described by the equity geography, options are: 'race', 'poverty', 'disability', 'elderly', 'limited_english'

year
string, the year for which the equity geography is defined

Value

a data frame with all tracts in the region with a 0-1 variable for whether they are in an equity geography or not

Author

Suzanne Childress

Examples

if (FALSE) {
Sys.getenv("CENSUS_API_KEY")
}

get_equity_geographies(equity_type = 'regavg', equity_group = 'disability', year = '2019')

#> Reading layer `OGRGeoJSON' from data source `https://services6.arcgis.com/GWxg6t7KXELn1thE/arcgis/rest/services/equity_tracts_2019/FeatureServer/0/query/?where=0=0&outFields=geoid,equity_geog_disabled_vs_reg_ave&f=pgeojson' using driver `GeoJSON'
#> Simple feature collection with 891 features and 2 fields
#> geometry type:  POLYGON
#> dimension:      XY
#> bbox:           xmin: -123.0234 ymin: 46.72799 xmax: -120.9062 ymax: 48.29924
#> geographic CRS: WGS 84

#>           geoid equity_geog_disabled_vs_reg_average
#> 1   53033022006                                   0
#> 2   53033032320                                   0
#> 3   53033031703                                   1
#> 4   53033031206                                   0
#> 5   53033032102                                   0
#> 6   53033032210                                   0
#> 7   53033032211                                   0
#> 8   53033032104                                   0
#> 9   53033032313                                   0
#> 10  53033032203                                   0
#> 11  53033032703                                   0
#> 12  53033032704                                   0
#> 13  53033027300                                   0
#> 14  53033027200                                   0
#> 15  53033027100                                   1
#> 16  53033027000                                   1
#> 17  53033026802                                   0
#> 18  53033026801                                   1
#> 19  53033026700                                   1
#> 20  53033026600                                   0
#> 21  53033000800                                   0
#> 22  53033000700                                   1
#> 23  53033000600                                   1
#> 24  53033000500                                   0
#> 25  53033000402                                   1
#> 26  53033000401                                   1
#> 27  53033000300                                   1
#> 28  53033000200                                   0
#> 29  53033000100                                   1
#> 30  53033032500                                   0
#> 31  53033031907                                   0
#> 32  53033031800                                   1
#> 33  53033032002                                   0
#> 34  53033032003                                   0
#> 35  53033031704                                   0
#> 36  53033032005                                   0
#> 37  53033032006                                   0
#> 38  53033032207                                   0
#> 39  53033021803                                   1
#> 40  53033032602                                   0
#> 41  53033032702                                   0
#> 42  53033032702                                   0
#> 43  53033032800                                   1
#> 44  53033026500                                   1
#> 45  53033026400                                   0
#> 46  53033026300                                   1
#> 47  53033026300                                   1
#> 48  53033026300                                   1
#> 49  53033026300                                   1
#> 50  53033026200                                   0
#> 51  53033026100                                   1
#> 52  53033026001                                   0
#> 53  53033021500                                   0
#> 54  53033021400                                   0
#> 55  53033021300                                   0
#> 56  53033001702                                   1
#> 57  53033004302                                   0
#> 58  53033004301                                   0
#> 59  53033007402                                   1
#> 60  53033010002                                   0
#> 61  53033010001                                   1
#> 62  53033011002                                   0
#> 63  53033010402                                   0
#> 64  53033010702                                   0
#> 65  53033011402                                   1
#> 66  53033022605                                   0
#> 67  53033022902                                   0
#> 68  53033024602                                   0
#> 69  53033025803                                   0
#> 70  53033031601                                   0
#> 71  53033031603                                   1
#> 72  53033032103                                   0
#> 73  53033025001                                   0
#> 74  53033032401                                   0
#> 75  53033032402                                   0
#> 76  53033031501                                   1
#> 77  53033021100                                   0
#> 78  53033021000                                   1
#> 79  53033020900                                   0
#> 80  53033020800                                   1
#> 81  53033020700                                   1
#> 82  53033020600                                   0
#> 83  53033020500                                   0
#> 84  53033011401                                   1
#> 85  53033025302                                   1
#> 86  53033025806                                   0
#> 87  53033027701                                   1
#> 88  53033032326                                   0
#> 89  53033032327                                   1
#> 90  53033023404                                   0
#> 91  53033023403                                   0
#> 92  53033023804                                   0
#> 93  53033023201                                   0
#> 94  53033023900                                   0
#> 95  53033032322                                   0
#> 96  53033032321                                   0
#> 97  53033021804                                   0
#> 98  53033022101                                   0
#> 99  53033022101                                   0
#> 100 53033025003                                   0
#> 101 53033022701                                   0
#> 102 53033032008                                   0
#> 103 53033032208                                   1
#> 104 53033020402                                   0
#> 105 53033020401                                   0
#> 106 53033020300                                   1
#> 107 53033020200                                   1
#> 108 53033020100                                   1
#> 109 53033012100                                   1
#> 110 53033023803                                   0
#> 111 53033025006                                   0
#> 112 53033025005                                   0
#> 113 53033025602                                   0
#> 114 53033029206                                   1
#> 115 53033029102                                   1
#> 116 53033029101                                   0
#> 117 53033030006                                   1
#> 118 53033030314                                   1
#> 119 53033030313                                   1
#> 120 53033031605                                   1
#> 121 53033024100                                   0
#> 122 53033022703                                   0
#> 123 53033024200                                   0
#> 124 53033022203                                   0
#> 125 53033024800                                   0
#> 126 53033024902                                   0
#> 127 53033024903                                   0
#> 128 53033025102                                   0
#> 129 53033031903                                   0
#> 130 53033031904                                   0
#> 131 53033031906                                   0
#> 132 53033031301                                   1
#> 133 53033031100                                   1
#> 134 53033031000                                   0
#> 135 53033030902                                   1
#> 136 53033030901                                   0
#> 137 53033030802                                   1
#> 138 53033030801                                   1
#> 139 53033012000                                   0
#> 140 53033011900                                   1
#> 141 53033010900                                   0
#> 142 53033010800                                   0
#> 143 53033010800                                   0
#> 144 53033010800                                   0
#> 145 53033010600                                   0
#> 146 53033010500                                   0
#> 147 53033010300                                   0
#> 148 53033031604                                   0
#> 149 53033032011                                   0
#> 150 53033031706                                   1
#> 151 53033032215                                   0
#> 152 53033032329                                   0
#> 153 53033032328                                   0
#> 154 53033001701                                   0
#> 155 53033007401                                   0
#> 156 53033032317                                   0
#> 157 53033032315                                   0
#> 158 53033032319                                   1
#> 159 53033021906                                   0
#> 160 53033021905                                   0
#> 161 53033032307                                   0
#> 162 53033032311                                   0
#> 163 53033032324                                   0
#> 164 53033022801                                   0
#> 165 53033030700                                   1
#> 166 53033030600                                   1
#> 167 53033030504                                   1
#> 168 53033030503                                   1
#> 169 53033030501                                   1
#> 170 53033030404                                   0
#> 171 53033010200                                   0
#> 172 53033010100                                   1
#> 173 53033009900                                   0
#> 174 53033009900                                   0
#> 175 53033009800                                   0
#> 176 53033009702                                   0
#> 177 53033009701                                   0
#> 178 53033009600                                   0
#> 179 53033010401                                   1
#> 180 53033010701                                   0
#> 181 53033011001                                   1
#> 182 53033025301                                   0
#> 183 53033025601                                   1
#> 184 53033025805                                   0
#> 185 53033022803                                   0
#> 186 53033023202                                   1
#> 187 53033031205                                   0
#> 188 53033031502                                   1
#> 189 53033022005                                   0
#> 190 53033021802                                   0
#> 191 53033021904                                   0
#> 192 53033030403                                   1
#> 193 53033030401                                   0
#> 194 53033030312                                   0
#> 195 53033030311                                   1
#> 196 53033030310                                   0
#> 197 53033009500                                   0
#> 198 53033009400                                   0
#> 199 53033009300                                   1
#> 200 53033009200                                   1
#> 201 53033009200                                   1
#> 202 53033009100                                   1
#> 203 53033009000                                   1
#> 204 53033008900                                   0
#> 205 53033008800                                   0
#> 206 53033029205                                   0
#> 207 53033030005                                   0
#> 208 53033031705                                   0
#> 209 53033032010                                   0
#> 210 53033032213                                   0
#> 211 53033032214                                   0
#> 212 53033022300                                   0
#> 213 53033022400                                   0
#> 214 53033022901                                   1
#> 215 53033022606                                   0
#> 216 53033022702                                   0
#> 217 53033023000                                   0
#> 218 53033030309                                   0
#> 219 53033030308                                   0
#> 220 53033030306                                   0
#> 221 53033030305                                   1
#> 222 53033008700                                   1
#> 223 53033011800                                   0
#> 224 53033011700                                   1
#> 225 53033011600                                   0
#> 226 53033011500                                   0
#> 227 53033011300                                   1
#> 228 53033011200                                   1
#> 229 53033011102                                   0
#> 230 53033032212                                   0
#> 231 53033023100                                   0
#> 232 53033023603                                   0
#> 233 53033023300                                   1
#> 234 53033023401                                   1
#> 235 53033023500                                   0
#> 236 53033030304                                   1
#> 237 53033030202                                   0
#> 238 53033030201                                   1
#> 239 53033030100                                   1
#> 240 53033030004                                   1
#> 241 53033030003                                   0
#> 242 53033011101                                   1
#> 243 53033008600                                   0
#> 244 53033008500                                   1
#> 245 53033008400                                   1
#> 246 53033008300                                   1
#> 247 53033008200                                   1
#> 248 53033008100                                   1
#> 249 53033008002                                   1
#> 250 53033008001                                   0
#> 251 53033007900                                   0
#> 252 53033007800                                   0
#> 253 53033007700                                   0
#> 254 53033023601                                   0
#> 255 53033024601                                   0
#> 256 53033023604                                   1
#> 257 53033023700                                   0
#> 258 53033029902                                   1
#> 259 53033029901                                   0
#> 260 53033029802                                   0
#> 261 53033029801                                   0
#> 262 53033029700                                   1
#> 263 53033029602                                   0
#> 264 53033029601                                   1
#> 265 53033029504                                   1
#> 266 53033007600                                   0
#> 267 53033007500                                   0
#> 268 53033007300                                   1
#> 269 53033007200                                   0
#> 270 53033007100                                   0
#> 271 53033007000                                   1
#> 272 53033006900                                   0
#> 273 53033006800                                   0
#> 274 53033006700                                   0
#> 275 53033006600                                   0
#> 276 53033006500                                   0
#> 277 53033023801                                   0
#> 278 53033024000                                   0
#> 279 53033024701                                   0
#> 280 53033024901                                   0
#> 281 53033024702                                   1
#> 282 53033021700                                   0
#> 283 53033022001                                   1
#> 284 53033029503                                   0
#> 285 53033029502                                   0
#> 286 53033029408                                   1
#> 287 53033029407                                   1
#> 288 53033029406                                   0
#> 289 53033029405                                   0
#> 290 53033006400                                   0
#> 291 53033006300                                   0
#> 292 53033006200                                   0
#> 293 53033006200                                   0
#> 294 53033006200                                   0
#> 295 53033006100                                   0
#> 296 53033006000                                   0
#> 297 53033005801                                   0
#> 298 53033005700                                   0
#> 299 53033026002                                   0
#> 300 53033025804                                   1
#> 301 53033031909                                   0
#> 302 53033031908                                   1
#> 303 53033022102                                   0
#> 304 53033022500                                   0
#> 305 53033029403                                   0
#> 306 53033029307                                   0
#> 307 53033029306                                   0
#> 308 53033028902                                   0
#> 309 53033028901                                   0
#> 310 53033028802                                   0
#> 311 53033028801                                   1
#> 312 53033005600                                   0
#> 313 53033005400                                   0
#> 314 53033005302                                   1
#> 315 53033005301                                   0
#> 316 53033005200                                   0
#> 317 53033005100                                   0
#> 318 53033005000                                   0
#> 319 53033004900                                   0
#> 320 53033004800                                   0
#> 321 53033004700                                   0
#> 322 53033022603                                   0
#> 323 53033022604                                   0
#> 324 53033024300                                   1
#> 325 53033024400                                   1
#> 326 53033024500                                   0
#> 327 53033028700                                   0
#> 328 53033028600                                   1
#> 329 53033028500                                   0
#> 330 53033028403                                   1
#> 331 53033028402                                   0
#> 332 53033028300                                   0
#> 333 53033004600                                   0
#> 334 53033004500                                   0
#> 335 53033004400                                   0
#> 336 53033004200                                   0
#> 337 53033004100                                   0
#> 338 53033004000                                   0
#> 339 53033003900                                   0
#> 340 53033003800                                   0
#> 341 53033025701                                   1
#> 342 53033022802                                   0
#> 343 53033032309                                   0
#> 344 53033032325                                   0
#> 345 53033028200                                   0
#> 346 53033028100                                   0
#> 347 53033028000                                   1
#> 348 53033027900                                   1
#> 349 53033027800                                   1
#> 350 53033027702                                   1
#> 351 53033005900                                   0
#> 352 53033005802                                   0
#> 353 53033003600                                   1
#> 354 53033003500                                   0
#> 355 53033003400                                   0
#> 356 53033003300                                   0
#> 357 53033003200                                   0
#> 358 53033003200                                   0
#> 359 53033003100                                   0
#> 360 53033025400                                   1
#> 361 53033025500                                   1
#> 362 53033025702                                   0
#> 363 53033032316                                   0
#> 364 53033032318                                   0
#> 365 53033031202                                   1
#> 366 53033031204                                   1
#> 367 53033031302                                   1
#> 368 53033032601                                   1
#> 369 53033027600                                   1
#> 370 53033029305                                   0
#> 371 53033029304                                   0
#> 372 53033029303                                   1
#> 373 53033029204                                   0
#> 374 53033003000                                   0
#> 375 53033002900                                   0
#> 376 53033002800                                   0
#> 377 53033002700                                   0
#> 378 53033002600                                   0
#> 379 53033002500                                   0
#> 380 53033002400                                   0
#> 381 53033002200                                   0
#> 382 53033002100                                   0
#> 383 53033002000                                   1
#> 384 53033031400                                   1
#> 385 53033032007                                   0
#> 386 53033025101                                   0
#> 387 53033025200                                   1
#> 388 53033022003                                   0
#> 389 53033022201                                   0
#> 390 53033029203                                   1
#> 391 53033029004                                   0
#> 392 53033029003                                   1
#> 393 53033029001                                   1
#> 394 53033027500                                   1
#> 395 53033027400                                   0
#> 396 53033001900                                   0
#> 397 53033001800                                   0
#> 398 53033001600                                   0
#> 399 53033001500                                   0
#> 400 53033001400                                   0
#> 401 53033001300                                   1
#> 402 53033001200                                   1
#> 403 53033001100                                   0
#> 404 53033001000                                   0
#> 405 53033000900                                   0
#> 406 53033022202                                   0
#> 407 53033032323                                   0
#> 408 53033021600                                   0
#> 409 53033021903                                   0
#> 410 53035090700                                   0
#> 411 53035090700                                   0
#> 412 53035090800                                   0
#> 413 53035090900                                   0
#> 414 53035091000                                   0
#> 415 53035080101                                   1
#> 416 53035080102                                   1
#> 417 53035080200                                   1
#> 418 53035080300                                   1
#> 419 53035080400                                   1
#> 420 53035080600                                   1
#> 421 53035080700                                   1
#> 422 53035092801                                   0
#> 423 53035091204                                   1
#> 424 53035091302                                   1
#> 425 53035091400                                   1
#> 426 53035091500                                   1
#> 427 53035091600                                   1
#> 428 53035091700                                   1
#> 429 53035091800                                   1
#> 430 53035092000                                   1
#> 431 53035092000                                   1
#> 432 53035080500                                   1
#> 433 53035080900                                   1
#> 434 53035081000                                   1
#> 435 53035081100                                   1
#> 436 53035081200                                   1
#> 437 53035081400                                   0
#> 438 53035091301                                   1
#> 439 53035091201                                   1
#> 440 53035091203                                   0
#> 441 53035091900                                   1
#> 442 53035091900                                   1
#> 443 53035090101                                   1
#> 444 53035090102                                   1
#> 445 53035092100                                   1
#> 446 53035092200                                   1
#> 447 53035092400                                   1
#> 448 53035092500                                   1
#> 449 53035092600                                   1
#> 450 53035092901                                   1
#> 451 53035092300                                   1
#> 452 53035090300                                   0
#> 453 53035090400                                   0
#> 454 53035940100                                   1
#> 455 53035940000                                   1
#> 456 53035940000                                   1
#> 457 53035091100                                   1
#> 458 53035080800                                   0
#> 459 53035092802                                   0
#> 460 53035092803                                   1
#> 461 53035092902                                   1
#> 462 53035090202                                   1
#> 463 53035090202                                   1
#> 464 53035090202                                   1
#> 465 53035090202                                   1
#> 466 53035090202                                   1
#> 467 53035090202                                   1
#> 468 53035090202                                   1
#> 469 53035090202                                   1
#> 470 53035090202                                   1
#> 471 53035090202                                   1
#> 472 53035090202                                   1
#> 473 53035090502                                   1
#> 474 53035090201                                   1
#> 475 53035090501                                   1
#> 476 53035092704                                   1
#> 477 53035092704                                   1
#> 478 53035092701                                   1
#> 479 53053062000                                   1
#> 480 53053071601                                   1
#> 481 53053073111                                   1
#> 482 53053940007                                   0
#> 483 53053940007                                   0
#> 484 53053940006                                   1
#> 485 53053073116                                   0
#> 486 53053072800                                   0
#> 487 53053072800                                   0
#> 488 53053072107                                   1
#> 489 53053071703                                   1
#> 490 53053073113                                   0
#> 491 53053072506                                   0
#> 492 53053071411                                   1
#> 493 53053071410                                   1
#> 494 53053072308                                   1
#> 495 53053940010                                   1
#> 496 53053940010                                   1
#> 497 53053940010                                   1
#> 498 53053940010                                   1
#> 499 53053071409                                   1
#> 500 53053073302                                   1
#> 501 53053073302                                   1
#> 502 53053071803                                   1
#> 503 53053073115                                   1
#> 504 53053072106                                   1
#> 505 53053072105                                   1
#> 506 53053062600                                   1
#> 507 53053073200                                   1
#> 508 53053073200                                   1
#> 509 53053073200                                   1
#> 510 53053073200                                   1
#> 511 53053073200                                   1
#> 512 53053073200                                   1
#> 513 53053073200                                   1
#> 514 53053072410                                   0
#> 515 53053072410                                   0
#> 516 53053073110                                   1
#> 517 53053073108                                   1
#> 518 53053071902                                   1
#> 519 53053072307                                   1
#> 520 53053072311                                   1
#> 521 53053072109                                   1
#> 522 53053072109                                   1
#> 523 53053072407                                   1
#> 524 53053073001                                   1
#> 525 53053070100                                   1
#> 526 53053070100                                   1
#> 527 53053073006                                   1
#> 528 53053070206                                   1
#> 529 53053071506                                   1
#> 530 53053071504                                   1
#> 531 53053070309                                   1
#> 532 53053070309                                   1
#> 533 53053063501                                   1
#> 534 53053071901                                   1
#> 535 53053062900                                   1
#> 536 53053063000                                   1
#> 537 53053070401                                   1
#> 538 53053063502                                   1
#> 539 53053072405                                   0
#> 540 53053072405                                   0
#> 541 53053061200                                   1
#> 542 53053073404                                   1
#> 543 53053073005                                   1
#> 544 53053072504                                   0
#> 545 53053072309                                   0
#> 546 53053072309                                   0
#> 547 53053060300                                   1
#> 548 53053071706                                   1
#> 549 53053071705                                   1
#> 550 53053071704                                   0
#> 551 53053072408                                   1
#> 552 53053071403                                   1
#> 553 53053070207                                   1
#> 554 53053071209                                   1
#> 555 53053072406                                   0
#> 556 53053061601                                   1
#> 557 53053061800                                   1
#> 558 53053072603                                   1
#> 559 53053072603                                   1
#> 560 53053072603                                   1
#> 561 53053072603                                   1
#> 562 53053072603                                   1
#> 563 53053072603                                   1
#> 564 53053072602                                   1
#> 565 53053072602                                   1
#> 566 53053072602                                   1
#> 567 53053072602                                   1
#> 568 53053072602                                   1
#> 569 53053072601                                   1
#> 570 53053072601                                   1
#> 571 53053061500                                   1
#> 572 53053072503                                   1
#> 573 53053061300                                   1
#> 574 53053061700                                   1
#> 575 53053070307                                   0
#> 576 53053061602                                   1
#> 577 53053070308                                   1
#> 578 53053072305                                   0
#> 579 53053062801                                   1
#> 580 53053071210                                   1
#> 581 53053071307                                   0
#> 582 53053062802                                   1
#> 583 53053063400                                   1
#> 584 53053072108                                   1
#> 585 53053071408                                   1
#> 586 53053073114                                   1
#> 587 53053071503                                   1
#> 588 53053061400                                   1
#> 589 53053940002                                   1
#> 590 53053072313                                   1
#> 591 53053071806                                   1
#> 592 53053062500                                   1
#> 593 53053072111                                   1
#> 594 53053072112                                   0
#> 595 53053940011                                   0
#> 596 53053071805                                   1
#> 597 53053060600                                   0
#> 598 53053061100                                   1
#> 599 53053072312                                   0
#> 600 53053071808                                   1
#> 601 53053071807                                   1
#> 602 53053073123                                   0
#> 603 53053070404                                   1
#> 604 53053070403                                   1
#> 605 53053070313                                   0
#> 606 53053070314                                   0
#> 607 53053070315                                   0
#> 608 53053070316                                   0
#> 609 53053073125                                   1
#> 610 53053072903                                   0
#> 611 53053071407                                   1
#> 612 53053060800                                   1
#> 613 53053073119                                   1
#> 614 53053072507                                   0
#> 615 53053072409                                   1
#> 616 53053071406                                   0
#> 617 53053073126                                   1
#> 618 53053073408                                   0
#> 619 53053073408                                   0
#> 620 53053073406                                   1
#> 621 53053073406                                   1
#> 622 53053073406                                   1
#> 623 53053073406                                   1
#> 624 53053073121                                   0
#> 625 53053071310                                   0
#> 626 53053071309                                   1
#> 627 53053072906                                   0
#> 628 53053940004                                   1
#> 629 53053940004                                   1
#> 630 53053940004                                   1
#> 631 53053071206                                   1
#> 632 53053062300                                   1
#> 633 53053060200                                   1
#> 634 53053060200                                   1
#> 635 53053060200                                   1
#> 636 53053060904                                   1
#> 637 53053061002                                   1
#> 638 53053061001                                   1
#> 639 53053070312                                   0
#> 640 53053940003                                   0
#> 641 53053073120                                   0
#> 642 53053073122                                   1
#> 643 53053073405                                   1
#> 644 53053073407                                   1
#> 645 53053073407                                   1
#> 646 53053073124                                   0
#> 647 53053070703                                   1
#> 648 53053071306                                   1
#> 649 53053073301                                   1
#> 650 53053060700                                   0
#> 651 53053072310                                   1
#> 652 53053072000                                   1
#> 653 53053073118                                   1
#> 654 53053072905                                   0
#> 655 53053072907                                   0
#> 656 53053073500                                   0
#> 657 53053061900                                   1
#> 658 53053070310                                   1
#> 659 53053072901                                   0
#> 660 53053071602                                   1
#> 661 53053071305                                   1
#> 662 53053073117                                   1
#> 663 53053072505                                   0
#> 664 53053940005                                   0
#> 665 53053060500                                   0
#> 666 53053060905                                   1
#> 667 53053060906                                   1
#> 668 53053060903                                   1
#> 669 53053060400                                   0
#> 670 53053063100                                   1
#> 671 53053070205                                   1
#> 672 53053071207                                   0
#> 673 53053063200                                   1
#> 674 53053071505                                   1
#> 675 53053070311                                   1
#> 676 53053940009                                   0
#> 677 53053940008                                   0
#> 678 53053070204                                   1
#> 679 53053071304                                   1
#> 680 53053071707                                   1
#> 681 53053070203                                   0
#> 682 53053071100                                   1
#> 683 53053940001                                   0
#> 684 53053071205                                   1
#> 685 53053071208                                   1
#> 686 53053062400                                   0
#> 687 53053063300                                   1
#> 688 53061041810                                   1
#> 689 53061051924                                   0
#> 690 53061051922                                   0
#> 691 53061051921                                   0
#> 692 53061051926                                   0
#> 693 53061052115                                   0
#> 694 53061052114                                   0
#> 695 53061052209                                   1
#> 696 53061052709                                   0
#> 697 53061053604                                   1
#> 698 53061053603                                   0
#> 699 53061052906                                   1
#> 700 53061053508                                   0
#> 701 53061052905                                   1
#> 702 53061041809                                   1
#> 703 53061051928                                   0
#> 704 53061041303                                   0
#> 705 53061051923                                   0
#> 706 53061051925                                   0
#> 707 53061052208                                   1
#> 708 53061052708                                   1
#> 709 53061053507                                   1
#> 710 53061053509                                   1
#> 711 53061040500                                   1
#> 712 53061052003                                   0
#> 713 53061052118                                   0
#> 714 53061041304                                   1
#> 715 53061990002                                   0
#> 716 53061990002                                   0
#> 717 53061990002                                   0
#> 718 53061990002                                   0
#> 719 53061990002                                   0
#> 720 53061990002                                   0
#> 721 53061990002                                   0
#> 722 53061990002                                   0
#> 723 53061990002                                   0
#> 724 53061990002                                   0
#> 725 53061990002                                   0
#> 726 53061990002                                   0
#> 727 53061990002                                   0
#> 728 53061990002                                   0
#> 729 53061990002                                   0
#> 730 53061041704                                   0
#> 731 53061051927                                   0
#> 732 53061041703                                   0
#> 733 53061041606                                   1
#> 734 53061040900                                   1
#> 735 53061041608                                   0
#> 736 53061041601                                   1
#> 737 53061041806                                   0
#> 738 53061041905                                   0
#> 739 53061041904                                   1
#> 740 53061041701                                   1
#> 741 53061041811                                   0
#> 742 53061052706                                   0
#> 743 53061052707                                   0
#> 744 53061051804                                   0
#> 745 53061051803                                   0
#> 746 53061041812                                   1
#> 747 53061053602                                   1
#> 748 53061040700                                   1
#> 749 53061040700                                   1
#> 750 53061041903                                   1
#> 751 53061052206                                   0
#> 752 53061041301                                   0
#> 753 53061053505                                   1
#> 754 53061053700                                   1
#> 755 53061042001                                   0
#> 756 53061042003                                   0
#> 757 53061050200                                   0
#> 758 53061050500                                   1
#> 759 53061050600                                   0
#> 760 53061050700                                   1
#> 761 53061051000                                   0
#> 762 53061041400                                   1
#> 763 53061041607                                   0
#> 764 53061041000                                   1
#> 765 53061051913                                   0
#> 766 53061051100                                   1
#> 767 53061051300                                   0
#> 768 53061051400                                   1
#> 769 53061051500                                   1
#> 770 53061051701                                   1
#> 771 53061051601                                   0
#> 772 53061052603                                   0
#> 773 53061052203                                   0
#> 774 53061051905                                   1
#> 775 53061051916                                   0
#> 776 53061052007                                   0
#> 777 53061051917                                   0
#> 778 53061052005                                   0
#> 779 53061050300                                   0
#> 780 53061050401                                   0
#> 781 53061050402                                   1
#> 782 53061050800                                   1
#> 783 53061051914                                   0
#> 784 53061052604                                   1
#> 785 53061052607                                   0
#> 786 53061052903                                   1
#> 787 53061052903                                   1
#> 788 53061053302                                   1
#> 789 53061053301                                   1
#> 790 53061053301                                   1
#> 791 53061053301                                   1
#> 792 53061053301                                   1
#> 793 53061053301                                   1
#> 794 53061053301                                   1
#> 795 53061053301                                   1
#> 796 53061053301                                   1
#> 797 53061053201                                   1
#> 798 53061053201                                   1
#> 799 53061053201                                   1
#> 800 53061053201                                   1
#> 801 53061053201                                   1
#> 802 53061053201                                   1
#> 803 53061053201                                   1
#> 804 53061053201                                   1
#> 805 53061053201                                   1
#> 806 53061053201                                   1
#> 807 53061053201                                   1
#> 808 53061053201                                   1
#> 809 53061053201                                   1
#> 810 53061053201                                   1
#> 811 53061053201                                   1
#> 812 53061053201                                   1
#> 813 53061053201                                   1
#> 814 53061053201                                   1
#> 815 53061053201                                   1
#> 816 53061050900                                   0
#> 817 53061042004                                   0
#> 818 53061051702                                   1
#> 819 53061051915                                   0
#> 820 53061051802                                   0
#> 821 53061052004                                   1
#> 822 53061051200                                   0
#> 823 53061042005                                   0
#> 824 53061052006                                   0
#> 825 53061040200                                   1
#> 826 53061040200                                   1
#> 827 53061040100                                   1
#> 828 53061040100                                   1
#> 829 53061040300                                   1
#> 830 53061040800                                   1
#> 831 53061040800                                   1
#> 832 53061040400                                   1
#> 833 53061041202                                   1
#> 834 53061041201                                   0
#> 835 53061041100                                   1
#> 836 53061041500                                   1
#> 837 53061041805                                   1
#> 838 53061051602                                   1
#> 839 53061042006                                   0
#> 840 53061052606                                   1
#> 841 53061052605                                   1
#> 842 53061051912                                   0
#> 843 53061052401                                   1
#> 844 53061052107                                   0
#> 845 53061052108                                   0
#> 846 53061052904                                   1
#> 847 53061041605                                   0
#> 848 53061052207                                   0
#> 849 53061052204                                   0
#> 850 53061053802                                   1
#> 851 53061041901                                   1
#> 852 53061052104                                   0
#> 853 53061052104                                   0
#> 854 53061052104                                   0
#> 855 53061052104                                   0
#> 856 53061052104                                   0
#> 857 53061052104                                   0
#> 858 53061052104                                   0
#> 859 53061052104                                   0
#> 860 53061052105                                   0
#> 861 53061052301                                   0
#> 862 53061052302                                   0
#> 863 53061052504                                   1
#> 864 53061052503                                   0
#> 865 53061052705                                   1
#> 866 53061052502                                   0
#> 867 53061053102                                   1
#> 868 53061053101                                   1
#> 869 53061053202                                   1
#> 870 53061053400                                   1
#> 871 53061053504                                   1
#> 872 53061052701                                   0
#> 873 53061052806                                   1
#> 874 53061052803                                   0
#> 875 53061052804                                   1
#> 876 53061940001                                   1
#> 877 53061940001                                   1
#> 878 53061940001                                   1
#> 879 53061940002                                   1
#> 880 53061940002                                   1
#> 881 53061053506                                   1
#> 882 53061053801                                   1
#> 883 53061053803                                   1
#> 884 53061050102                                   0
#> 885 53061050101                                   1
#> 886 53061051918                                   0
#> 887 53061052112                                   1
#> 888 53061052113                                   0
#> 889 53061052402                                   1
#> 890 53061052805                                   1
#> 891 53061041808                                   1