Get equity tracts

Retrieve a 0-1 variable for whether each tract in the region is in an equity geography or not

get_equity_geographies(
  elmer_connection,
  equity_type = "equity_geog_vs_50_percent",
  equity_group = "racial_equity_geographies",
  year = "2016"
)

Arguments

elmer_connection
A DBI connection object created from the library odbc

equity_type
string, The type of equity geography either 'equity_geog_vs_50_percent' or 'equity_geog_vs_reg_total'

equity_group
string, The group of people described by the equity geography, options are:'racial_equity_geographies', 'poverty_equity_geographies', 'disability_equity_geographies', 'elderly_equity_geographies', 'limited_english_equity_geographies'

year
string, the year for which the equity geography is defined

Value

a data frame with all tracts in the region with a 0-1 variable for whether they are in an equity geography or not

Author

Suzanne Childress

Examples

library(odbc)
library(DBI)
library(dplyr)
if (FALSE) {
Sys.getenv("CENSUS_API_KEY")
}
elmer_connection <- DBI::dbConnect(odbc::odbc(),
                             driver = "SQL Server",
                             server = "AWS-PROD-SQL\\Sockeye",
                             database = "Elmer",
                             trusted_connection = "yes")

poverty_tracts<-get_equity_geographies(elmer_connection, 'equity_geog_vs_reg_total', 'poverty_equity_geographies','2018')

#> Error in get_equity_geographies(elmer_connection, "equity_geog_vs_reg_total",     "poverty_equity_geographies", "2018"): could not find function "get_equity_geographies"

means_transport<-get_acs_recs('tract',table.names=c('S0801_C01'),   years=c(2019),
                             acs.type = 'acs1')

#> Getting data from the 2015-2019 5-year ACS

#> Loading ACS5/SUBJECT variables for 2019 from table S0801_C01. To cache this dataset for faster access to ACS tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per ACS dataset.

#> Using the ACS Subject Tables

#> Using the ACS Subject Tables

#> Using the ACS Subject Tables

tot_workers<-means_transport%>%filter(variable=='S0801_C01_001')
transit<-means_transport%>%filter(variable=='S0801_C01_009')
poverty_wrkrs<-merge(poverty_tracts, tot_workers, by.x='geoid', by.y='GEOID')

#> Error in merge(poverty_tracts, tot_workers, by.x = "geoid", by.y = "GEOID"): object 'poverty_tracts' not found

poverty_transit<-merge(poverty_wrkrs, transit, by.x='geoid', by.y='GEOID')

#> Error in merge(poverty_wrkrs, transit, by.x = "geoid", by.y = "GEOID"): object 'poverty_wrkrs' not found

poverty_transit_calcs<-poverty_transit%>%mutate(transit=estimate.x*estimate.y/100)%>%
mutate(workers=estimate.x)%>%select('geoid', 'transit', 'workers', 'equity_geog_vs_reg_total')

#> Error in mutate(., transit = estimate.x * estimate.y/100): object 'poverty_transit' not found

poverty_transit_summary<-poverty_transit_calcs%>%group_by(equity_geog_vs_reg_total)%>%
summarise(sum_workers=sum(workers), sum_transit=sum(transit, na.rm=TRUE))%>%
mutate(transit_share=sum_transit/sum_workers)

#> Error in group_by(., equity_geog_vs_reg_total): object 'poverty_transit_calcs' not found

poverty_transit_summary

#> Error in eval(expr, envir, enclos): object 'poverty_transit_summary' not found