Group ACS or Census Data

Group ACS to Census Data into pre-defined commonly used groupings, stored in the file: variables_groupings.

group_recs(tbl, variables_groupings, group_name)

Arguments

tbl
A data frame of census or acs data

variables_groupings
A data frame that provides look-ups between the variables and their categories

group_name
A character string that describes the grouping of the data. Should match the group_name in the variables_cats table

Value

a tibble of grouped ACS or Census estimates

Author

Suzanne Childress

Examples

if (FALSE) {
Sys.getenv("CENSUS_API_KEY")}
variables_groupings<-system.file('extdata', 'variables_groupings.csv', package='psrccensus')
age_sex_acs<-get_acs_recs(geography = 'county',
                     table.names = c('B01001'),
                     years=c(2019),
                     acs.type = 'acs1')

#> The 1-year ACS provides data for geographies with populations of 65,000 and greater.

#> Getting data from the 2019 1-year ACS

sex_age_census<-get_decennial_recs(geography = 'county',
                                  table_codes = c('P012'),
                                  year=2010)

#> Getting data from the 2010 decennial Census

#> Loading SF1 variables for 2010 from table P012. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.

#> Using Census Summary File 1

#> Using Census Summary File 1

#> Getting data from the 2010 decennial Census

#> Loading SF1 variables for 2010 from table P012. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.

#> Using Census Summary File 1

#> Using Census Summary File 1

#> Getting data from the 2010 decennial Census

#> Loading SF1 variables for 2010 from table P012. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.

#> Using Census Summary File 1

#> Using Census Summary File 1

#> Getting data from the 2010 decennial Census

#> Loading SF1 variables for 2010 from table P012. To cache this dataset for faster access to Census tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per Census dataset.

#> Using Census Summary File 1

#> Using Census Summary File 1

group_recs(sex_age_census, variables_groupings,'Broad Age')

#> # A tibble: 4 x 5
#> # Groups:   GEOID, NAME, concept [4]
#>   GEOID NAME                         concept    y                          value
#>   <chr> <chr>                        <chr>      <chr>                      <dbl>
#> 1 53033 King County, Washington      SEX BY AGE C:/Users/SChildress/Docu~ 5.79e6
#> 2 53035 Kitsap County, Washington    SEX BY AGE C:/Users/SChildress/Docu~ 7.53e5
#> 3 53053 Pierce County, Washington    SEX BY AGE C:/Users/SChildress/Docu~ 2.39e6
#> 4 53061 Snohomish County, Washington SEX BY AGE C:/Users/SChildress/Docu~ 2.14e6

income_acs<-get_acs_recs(geography = 'county',
table.names = c('B19001'),
years=c(2019),
acs.type = 'acs1')

#> The 1-year ACS provides data for geographies with populations of 65,000 and greater.

#> Getting data from the 2019 1-year ACS

group_recs(income_acs, variables_groupings,'Broad Income')

#> # A tibble: 5 x 10
#> # Groups:   GEOID, name, state, concept, census_geography, acs_type, year [5]
#>   GEOID  name   state concept   census_geography acs_type  year y       estimate
#>   <chr>  <chr>  <chr> <chr>     <chr>            <chr>    <dbl> <chr>      <dbl>
#> 1 53033  King ~ Wash~ HOUSEHOL~ County           acs1      2019 C:/Use~  1815522
#> 2 53035  Kitsa~ Wash~ HOUSEHOL~ County           acs1      2019 C:/Use~   215050
#> 3 53053  Pierc~ Wash~ HOUSEHOL~ County           acs1      2019 C:/Use~   667314
#> 4 53061  Snoho~ Wash~ HOUSEHOL~ County           acs1      2019 C:/Use~   599654
#> 5 REGION Region Wash~ HOUSEHOL~ County           acs1      2019 C:/Use~  3297540
#> # ... with 1 more variable: moe <dbl>